# backend/backend/docs/swagger/messenger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Suda Messenger: authentication, chats, messaging, users, search.",
        "title": "Suda Messenger Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/user/avatar": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Client uploads avatar via media-service first, then passes media_id here",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user avatar by media ID",
                "parameters": [
                    {
                        "description": "Avatar media ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_user.UpdateAvatarReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/device": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Register a device for push notifications",
                "parameters": [
                    {
                        "description": "Device data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_user.RegisterDeviceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/id/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user profile by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_user.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/id/{id}/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user online status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_user.UserStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/init-data": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Generate init data for mini-app integration",
                "responses": {
                    "200": {
                        "description": "initData string",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Log out and revoke session",
                "parameters": [
                    {
                        "description": "Logout data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_user.LogoutReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get current user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_user.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Verifies old password, sets new one, revokes all refresh sessions (user re-logs in everywhere).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Change password (authenticated)",
                "parameters": [
                    {
                        "description": "Old and new password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_user.ChangePasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond"
                        }
                    },
                    "400": {
                        "description": "INVALID_OLD_PASSWORD | VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/profile": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "description": "Profile data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_user.UpdateProfileReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "List active refresh sessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_features_user.SessionInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/sessions/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "Revoke a specific refresh session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_moera-sudo_backend_backend_messenger-service_internal_features_user.RegisterDeviceReq": {
            "type": "object",
            "required": [
                "fcm_token"
            ],
            "properties": {
                "device_name": {
                    "type": "string"
                },
                "fcm_token": {
                    "type": "string"
                }
            }
        },
...
        "github_com_moera-sudo_backend_backend_messenger-service_internal_features_user.UpdateAvatarReq": {
            "type": "object",
            "required": [
                "media_id"
            ],
            "properties": {
                "media_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_moera-sudo_backend_backend_messenger-service_internal_pkg_http.MessageRespond": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}